tests: Small tweaks for local iteration
Support e.g. `-e tests=payload-link`, to choose specific tests for more rapid
iteration, and allow skipping tmpdir cleanup to be able to debug.

+++ tests/installed/libinsttest.sh
@@ -25,7 +25,8 @@ dn=$(dirname $0)
 # Copy of bits from tap-test
 test_tmpdir=
 function _tmpdir_cleanup () {
-    if test -n "${test_tmpdir}" && test -f ${test_tmpdir}/.testtmp; then
+    if test -z "${TEST_SKIP_CLEANUP:-}" &&
+       test -n "${test_tmpdir}" && test -f ${test_tmpdir}/.testtmp; then
 	      rm "${test_tmpdir}" -rf
     fi
 }
